pams-chat-agent/requirements.txt
# =========================
# Core API
# =========================
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
flask==3.0.3

# =========================
# Database
# =========================
psycopg2-binary==2.9.9

# =========================
# Vector / RAG utilities
# =========================
numpy==1.26.4
scikit-learn==1.5.0
sentence-transformers>=2.2.0

# =========================
# PDF processing
# =========================
pymupdf==1.24.7
Pillow==10.3.0
pytesseract>=0.3.10
opencv-python>=4.9.0
img2table>=1.0.0
paddleocr>=2.7.0

# =========================
# Data processing
# =========================
pandas>=2.0
lxml==5.2.2
beautifulsoup4==4.12.3
html5lib==1.1

# =========================
# SQL parsing
# =========================
sqlglot>=23.0.0

# =========================
# HTTP client 
# =========================
httpx==0.27.0
requests==2.32.3

# =========================
# Utilities
# =========================
tqdm==4.66.4
loguru==0.7.2


# =========================
# LangChain + MCP Agents
# =========================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
mcp>=1.0.0
